Fix build break

+++ sql-plugin/src/main/java/com/nvidia/spark/rapids/shuffle/kudo/KudoSerializer.java
@@ -78,7 +78,7 @@ public class KudoSerializer {
     try {
       DataWriter writer = writerFrom(out);
       SerializedTableHeader header = new SerializedTableHeader(0, safeLongToInt(numRows), 0, 0, 0
-          , 0, new long[0]);
+          , 0, new byte[0]);
       header.writeTo(writer);
       writer.flush();
       return header.getSerializedSize();
